ocr-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
easyocr==1.7.1
pydantic-settings==2.3.3
# EasyOCR 运行依赖（图像与矩阵库）
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0
